# docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST para gerenciamento de eventos e pontuações da gamificação DNJ.",
        "title": "DNJ Game API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/subscriptions/webhook": {
            "post": {
                "description": "Grava o payload bruto e cria/atualiza um código de verificação pendente para o inscrito. Payload elástico (formato definido pela plataforma externa).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Recebe o webhook de inscrição no evento",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Secret compartilhado com a plataforma externa",
                        "name": "X-Webhook-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Payload do webhook (formato elástico)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Payload recebido"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_dnjtechteam_dnj-game-api_internal_app_errors.Error"
                        }
                    },
                    "401": {
                        "description": "Secret inválido"
                    }
                }
            }
        },
...
        "/users/{id}/update-group": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Vincula o usuário a um grupo existente (groupId) ou cria um novo grupo pelo nome (groupName) e vincula.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Atualiza o grupo de um usuário",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID do usuário",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Grupo a vincular",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_dnjtechteam_dnj-game-api_internal_app_messages.UpdateUserGroupRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dnjtechteam_dnj-game-api_internal_app_messages.UserResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_dnjtechteam_dnj-game-api_internal_app_errors.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dnjtechteam_dnj-game-api_internal_app_errors.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_dnjtechteam_dnj-game-api_internal_app_errors.Error": {
            "type": "object",
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_dnjtechteam_dnj-game-api_internal_app_errors.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_dnjtechteam_dnj-game-api_internal_app_errors.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_dnjtechteam_dnj-game-api_internal_app_messages.GroupSummaryDTO": {
            "type": "object",
            "properties": {
                "groupName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_dnjtechteam_dnj-game-api_internal_app_messages.UpdateUserGroupRequestDTO": {
            "type": "object",
            "properties": {
                "groupId": {
                    "type": "integer"
                },
                "groupName": {
                    "type": "string"
                }
            }
        },
        "github_com_dnjtechteam_dnj-game-api_internal_app_messages.UserResponseDTO": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "document": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "group": {
                    "$ref": "#/definitions/github_com_dnjtechteam_dnj-game-api_internal_app_messages.GroupSummaryDTO"
                },
                "id": {
                    "type": "integer"
                },
                "mobilePhone": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Identity token (JWT). Também aceito via cookie identity_token; o header é a forma documentada aqui.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}